Departament/requirements.txt
ansicon==1.89.0
asgiref==3.8.1
automium==0.2.6
automium_web==0.1.1
blessed==1.20.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Django==5.1.7
django-command==2.2.0
django-debug-toolbar==5.0.1
django-js-asset==3.1.2
django-mptt==0.16.0
editor==1.6.6
Employee==0.0.4
Faker==37.0.0
Flask==3.1.0
inquirer==3.4.0
itsdangerous==2.2.0
Jinja2==3.1.6
jinxed==1.3.0
legacy==0.1.7
MarkupSafe==3.0.2
python-dateutil==2.9.0.post0
readchar==4.2.1
runs==1.2.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
wcwidth==0.2.13
Werkzeug==3.1.3
xmod==1.8.1
